Noshooting
Added it so the ball can't be shot when moving. Also from here on out I'll try to follow the tutorial less to just see if I can. What I want to implement is holes, the potion collecting part, dif requirements for each level, and maybe a title screen, stroke and stroke counter.